gem_vertex/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
google-genai>=1.10.0
google-cloud-texttospeech>=2.16.0
google-auth>=2.28.0
sounddevice==0.5.1
numpy==2.1.3
mediapipe==0.10.14
opencv-python==4.10.0.84
chromadb==0.5.23
websockets==14.1
